telegram_bot/requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.9
filelock>=3.12
